requirements.txt
apscheduler
deepface
fastapi
groq
noobstuffs
pydantic
PyPDF2
python-docx
python-dotenv
python-multipart
tf-keras
uvicorn
web3
